Finished exercises 4.15 - 4.23

+++ bloglist-backend/controllers/blogs.js
@@ -1,30 +1,61 @@
+/* eslint-disable no-underscore-dangle */
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('creator', { username: 1, name: 1 });
   response.json(blogs);
 });
 
 blogsRouter.post('/', async (request, response) => {
-  const newBlog = request.body;
-  if (newBlog.title && newBlog.url) {
-    newBlog.likes = newBlog.likes || 0;
-    const newBlogObject = new Blog(newBlog);
-    const result = await newBlogObject.save();
-    response.status(201).json(result);
+  const {
+    title, author, url, likes,
+  } = request.body;
+  const { user } = request;
+
+  const missingTitle = !title;
+  const missingUrl = !url;
+  const errorMessage = [];
+
+  if (missingTitle) {
+    errorMessage.push('title is required.');
+  }
+  if (missingUrl) {
+    errorMessage.push('url is required.');
+  }
+  if (errorMessage.length > 0) {
+    response.status(400).json({
+      error: errorMessage.join(' '),
+    });
   } else {
-    response.status(400).end();
+    const newBlog = {
+      title,
+      author,
+      url,
+      likes: likes || 0,
+      creator: user._id,
+    };
+    const newBlogObject = new Blog(newBlog);
+    const savedBlog = await newBlogObject.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
+    response.status(201).json(savedBlog);
   }
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndRemove(request.params.id);
-  if (result) {
-    response.status(204).end();
-  } else {
-    response.status(404).end();
+  const { user } = request;
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(404).json({ error: 'blog that matches the id not found' });
   }
+  if (blog.creator.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'user unauthorized to delete blog' });
+  }
+
+  await Blog.findByIdAndRemove(request.params.id);
+  return response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
